disk index: index files grow by bytes instead of doubling

+++ bucket_map/src/bucket_map.rs
@@ -47,12 +47,15 @@ impl<T: Clone + Copy + Debug> std::fmt::Debug for BucketMap<T> {
     }
 }
 
+/// used to communicate resize necessary and current size.
 #[derive(Debug)]
 pub enum BucketMapError {
     /// (bucket_index, current_capacity_pow2)
+    /// Note that this is specific to data buckets, which grow in powers of 2
     DataNoSpace((u64, u8)),
-    /// current_capacity_pow2
-    IndexNoSpace(u8),
+    /// current_capacity_entries
+    /// Note that this is specific to index buckets, which can be 'random' sizes
+    IndexNoSpace(u64),
 }
 
 impl<T: Clone + Copy + Debug> BucketMap<T> {
